// web/src/gen/types/Link.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Link = {
  /**
   * @type string | undefined
   */
  label?: string
  /**
   * @type integer | undefined
   */
  order?: number
  /**
   * @type string | undefined
   */
  type?: string
  /**
   * @type string | undefined
   */
  url?: string
}